refactor(summary-notification): notifierインターフェースのファイル名修正
