+++ lambda/summary-notification/src/application/__init__.py
@@ -1,4 +1,4 @@
-from .notifier import INotifier, NotificationFailed
+from .notifier_interface import INotifier, NotificationFailed
 from .notify_weekly_summary_interface import INotifyWeeklySummaryUseCase
 from .notify_weekly_summary_usecase import NotifyWeeklySummaryUseCase
