+++ lambda/summary-notification/src/application/notifier_interface.py
